Move twemproxy config types to resource_builders/twemproxy pkg
