+++ pkg/generators/twemproxyconfig/configmap_test.go
@@ -1,13 +1,12 @@
 package twemproxyconfig
 
 import (
-	"reflect"
 	"testing"
 
 	saasv1alpha1 "github.com/3scale/saas-operator/api/v1alpha1"
 	"github.com/3scale/saas-operator/pkg/generators"
+	"github.com/3scale/saas-operator/pkg/resource_builders/twemproxy"
 	"github.com/go-test/deep"
-	"github.com/google/go-cmp/cmp"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
@@ -16,8 +15,8 @@ func TestGenerator_configMap(t *testing.T) {
 	type fields struct {
 		BaseOptionsV2  generators.BaseOptionsV2
 		Spec           saasv1alpha1.TwemproxyConfigSpec
-		masterTargets  map[string]TwemproxyServer
-		slaverwTargets map[string]TwemproxyServer
+		masterTargets  map[string]twemproxy.Server
+		slaverwTargets map[string]twemproxy.Server
 	}
 	type args struct {
 		toYAML bool
@@ -68,11 +67,11 @@ func TestGenerator_configMap(t *testing.T) {
 						},
 					},
 				},
-				masterTargets: map[string]TwemproxyServer{
+				masterTargets: map[string]twemproxy.Server{
 					"pshard01": {Address: "127.0.0.1:6379", Priority: 1, Name: "pshard01"},
 					"pshard02": {Address: "127.0.0.2:6379", Priority: 1, Name: "pshard02"},
 				},
-				slaverwTargets: map[string]TwemproxyServer{},
+				slaverwTargets: map[string]twemproxy.Server{},
 			},
 			args: args{},
 			want: &corev1.ConfigMap{
@@ -114,11 +113,11 @@ func TestGenerator_configMap(t *testing.T) {
 						},
 					},
 				},
-				masterTargets: map[string]TwemproxyServer{
+				masterTargets: map[string]twemproxy.Server{
 					"pshard01": {Address: "127.0.0.1:6379", Priority: 1, Name: "pshard01"},
 					"pshard02": {Address: "127.0.0.2:6379", Priority: 1, Name: "pshard02"},
 				},
-				slaverwTargets: map[string]TwemproxyServer{
+				slaverwTargets: map[string]twemproxy.Server{
 					"pshard01": {Address: "127.0.0.3:6379", Priority: 1, Name: "pshard01"},
 					"pshard02": {Address: "127.0.0.4:6379", Priority: 1, Name: "pshard02"},
 				},
@@ -148,98 +147,6 @@ func TestGenerator_configMap(t *testing.T) {
 			if diff := deep.Equal(got, tt.want); len(diff) > 0 {
 				t.Errorf("Generator.configMap() = diff %v", diff)
 			}
-		})
-	}
-}
-
-func TestTwemproxyServer_MarshalJSON(t *testing.T) {
-	type fields struct {
-		Address  string
-		Priority int
-		Name     string
-	}
-	tests := []struct {
-		name    string
-		fields  fields
-		want    []byte
-		wantErr bool
-	}{
-		{
-			name: "marshal",
-			fields: fields{
-				Address:  "127.0.0.1:6379",
-				Priority: 1,
-				Name:     "server",
-			},
-			want:    []byte("\"127.0.0.1:6379:1 server\""),
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tserver := &TwemproxyServer{
-				Address:  tt.fields.Address,
-				Priority: tt.fields.Priority,
-				Name:     tt.fields.Name,
-			}
-			got, err := tserver.MarshalJSON()
-			if (err != nil) != tt.wantErr {
-				t.Errorf("TwemproxyServer.MarshalJSON() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("TwemproxyServer.MarshalJSON() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestTwemproxyServer_UnmarshalJSON(t *testing.T) {
-	type args struct {
-		data []byte
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    *TwemproxyServer
-		wantErr bool
-	}{
-		{
-			name: "unmarshal",
-			args: args{
-				data: []byte("\"127.0.0.1:6379:2 server\""),
-			},
-			want: &TwemproxyServer{
-				Address:  "127.0.0.1:6379",
-				Priority: 2,
-				Name:     "server",
-			},
-			wantErr: false,
-		},
-		{
-			name: "error",
-			args: args{
-				data: []byte("\"127.0.0.1:6379:x server\""),
-			},
-			want:    nil,
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tserver := &TwemproxyServer{}
-			err := tserver.UnmarshalJSON(tt.args.data)
-			if err != nil {
-				if tt.wantErr != true {
-					t.Fatalf("TwemproxyServer.UnmarshalJSON() error = %v, wantErr %v", err, tt.wantErr)
-				}
-				return
-			} else {
-				if diff := cmp.Diff(tserver, tt.want); len(diff) != 0 {
-					t.Fatalf("TwemproxyServer.UnmarshalJSON() diff = %v", diff)
-				}
-			}
-
 		})
 	}
 }
